+++ lib/rib/test.rb
@@ -85,3 +85,8 @@ module Kernel
     self == rhs
   end
 end
+
+Rib::Blackhole = Object.new
+b = Rib::Blackhole.singleton_class
+b.instance_methods(true).each{ |m|
+  b.send(:undef_method, m) unless [:object_id, :__send__].include?(m) }